chore: new folder for permissions hooks
